+++ Assignment 3/VS_fesslerl_Chat/app/src/main/java/ch/ethz/inf/vs/a3/udpclient/ChatActivity.java
@@ -62,188 +62,12 @@ public class ChatActivity extends AppCompatActivity {
     @Override
     protected void onStart(){
         super.onStart();
-        connectToServer();
     }
 
     // Implement connection tear-down here.
     @Override
     protected void onStop(){
         super.onStop();
-        disconnectFromServer();
-    }
-
-    // TODO: Execute this code in an AsyncTask
-    // Called to connect to server (by sending a registration packet)
-    private void connectToServer(){
-
-        port = getPortFromSettings();
-
-        // Create new UDP Socket
-        try {
-            sock = new DatagramSocket(port);
-        }catch(SocketException e){
-            errorDiag("Could not bind socket to port " + port);
-            e.printStackTrace();
-        }
-
-        try {
-            sock.setSoTimeout(NetworkConsts.SOCKET_TIMEOUT);
-        } catch (SocketException e) {
-            errorDiag("Could not set socket timeout " + NetworkConsts.SOCKET_TIMEOUT);
-        }
-
-        toAddr = getIpFromSettings();
-
-
-        // Create Register Packet message
-        byte[] buf = new byte[0];
-        try {
-            buf = new Message(uname, uuid, null, MessageTypes.REGISTER, null).json.getBytes();
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-
-        // Build packet
-        DatagramPacket registerPacket = new DatagramPacket(buf, 0, buf.length, toAddr, port);
-
-        // Create new packet for ack message
-        byte[] ack_buf = new byte[256];     // More than enough space for the ack message
-        DatagramPacket ack = new DatagramPacket(ack_buf, ack_buf.length);
-
-        try {
-            sock.send(registerPacket);
-        } catch (IOException e) {
-            errorDiag("Could not send registration message.");
-        }
-
-
-        // TODO: Repeat the receive operation if timeout is reached (max. 5 times).
-        try {
-            sock.receive(ack);
-        } catch (SocketTimeoutException te) {
-            errorDiag("Socket timeout reached. Retrying...");
-        } catch (IOException e) {
-            errorDiag("Could not receive registration ack.");
-        }
-
-        Log.d("RECEIVED", "### " + ack.getData().toString());
-
-
-    }
-
-    // TODO: Could use JSONObject instead of hardcoded strings
-    // This method creates the message (json) for the register packet
-    /*private byte[] createRegisterMessage(){
-
-        StringBuilder sb = new StringBuilder();
-
-        // Create registration message line by line
-        sb.append("{\n");
-        sb.append("\"header\": {\n");
-        sb.append("\"username\": \"").append(uname).append("\",\n");
-        sb.append("\"uuid\": \"").append(uuid).append("\",\n");
-        sb.append("\"timestamp\": ").append("\"{}\",\n");
-        sb.append("\"type\": ").append("\"register\"\n");
-        sb.append("},\n");
-        sb.append("\"body\": {}\n");
-        sb.append("}");
-
-
-        return sb.toString().getBytes();
-    }*/
-
-    // Called to disconnect from server
-    private void disconnectFromServer(){
-
-        // TODO: Check if it is necessary to get port and ip value from the settings again
-
-        // Check if socket is null.
-        if(sock == null){
-            errorDiag("Socket null");
-        }
-
-        // Create packet
-        byte[] buf = new byte[0];
-        try {
-            buf = new Message(uname, uuid, null, MessageTypes.DEREGISTER, null).json.getBytes();
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-
-        DatagramPacket deregisterPacket = new DatagramPacket(buf, buf.length, toAddr, port);
-
-        // Create new packet for ack message
-        byte[] ack_buf = new byte[256];     // More than enough space for the ack message
-        DatagramPacket ack = new DatagramPacket(ack_buf, ack_buf.length);
-
-        try {
-            sock.send(deregisterPacket);
-        } catch(IOException e) {
-            errorDiag("Could not send deregistration message");
-        }
-
-        // TODO: ?? Repeat this receive operation until timeout is reached (max. 5 times). Not sure if necessary.
-        try {
-            sock.receive(ack);
-        } catch(SocketTimeoutException te){
-            errorDiag("Socket timeout reached. Retrying...");
-        } catch (IOException e) {
-            errorDiag("Could not receive deregistration ack.");
-        }
-
-        Log.d(LOG_TAG, "### " + ack.getData().toString());
-
-    }
-
-    // TODO: Could use JSONObject instead of hardcoded strings
-    // This method creates the message (jsoon) for the deregister packet
-    /*private byte[] createDeregisterMessage(){
-
-        StringBuilder sb = new StringBuilder();
-
-        // Create deregistrateion message line by line
-        sb.append("{\n");
-        sb.append("\"header\": {\n");
-        sb.append("\"username\": \"").append(uname).append("\",\n");
-        sb.append("\"uuid\": \"").append(uuid).append("\",\n");
-        sb.append("\"timestamp\": ").append("\"{}\",\n");
-        sb.append("\"type\": ").append("\"deregister\"\n");
-        sb.append("},\n");
-        sb.append("\"body\": {}\n");
-        sb.append("}");
-
-        return sb.toString().getBytes();
-    }*/
-
-    // TODO: Actually get the port from settings.
-    private int getPortFromSettings(){
-
-        int result = NetworkConsts.UDP_PORT;
-
-        // TODO: Move this part of error handling to SettingsActivity class.
-        // Error if port value is too small or too big
-        if (result < 1025 || result > 65535){
-            Log.d(LOG_TAG, "#### ERROR: Port value not in bounds.");
-            errorDiag(getString(R.string.error_port));
-            return NetworkConsts.UDP_PORT;
-        }
-
-        return result;
-    }
-
-    // TODO: Actually get the IP from settings.
-    private InetAddress getIpFromSettings(){
-
-        String ipString = NetworkConsts.SERVER_ADDRESS;
-
-        InetAddress addr = null;
-        try {
-            addr = InetAddress.getByName(ipString);
-        } catch (UnknownHostException e) {
-            errorDiag("Could not get IP Address by name " + ipString);
-        }
-
-        return addr;
     }
 
     // TODO: Implement dialog creation for displaying error messages
